update-player-logic
